assigment2/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
joblib==1.4.2
keras==3.9.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.5.5.62
opt-einsum==3.3.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
threadpoolctl==3.6.0
typing-extensions==3.7.4.3
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.12.1
zipp==3.21.0
